chore: rename jellyseerrClient to just seerrClient
