+++ src/pages/LibraryPage.js
@@ -20,7 +20,7 @@ import { i18n } from '../utils/i18n.js';
 import { state } from '../core/StateManager.js';
 import { storage } from '../utils/StorageService.js';
 import { eventBus } from '../core/EventBus.js';
-import { seerr } from '../api/JellyseerrClient.js';
+import { seerr } from '../api/seerrClient.js';
 import {
     getGenreNameById,
     getGenreMediaType,
