+++ src/pages/PersonPage.js
@@ -17,7 +17,7 @@ import { i18n } from '../utils/i18n.js';
 import { state } from '../core/StateManager.js';
 
 import FavoriteButton from '../components/FavoriteButton.js';
-import { seerr } from '../api/JellyseerrClient.js';
+import { seerr } from '../api/seerrClient.js';
 import DescriptionModal from '../components/DescriptionModal.js';
 import BackdropManager from '../utils/BackdropManager.js';
 import CardRenderer from '../utils/CardRenderer.js';
